move protocol ad child module of `process`, where it belongs
